# NAS-109081 / 21.02 / Make retrieving single chart release faster