This commit introduces changes to make retrieving single chart release considerably faster. With these changes we get chart release details in about 0.7 seconds and without this we get it in around 2 seconds for about 25 chart releases. And when chart release chart schema is required, it is 0.85 secs and the latter around 4.8 seconds.